fix: safe (or other sca) gas estimates
