+++ src/components/@molecules/TransactionDialogManager/stage/query.ts
@@ -1,17 +1,8 @@
 import { QueryFunctionContext } from '@tanstack/react-query'
 import { CallParameters, getFeeHistory, SendTransactionReturnType } from '@wagmi/core'
 import { Dispatch } from 'react'
-import {
-  Address,
-  BlockTag,
-  Hash,
-  Hex,
-  PrepareTransactionRequestRequest,
-  toHex,
-  Transaction,
-  TransactionRequest,
-} from 'viem'
-import { call, estimateGas, getTransaction, prepareTransactionRequest } from 'viem/actions'
+import { Hash, PrepareTransactionRequestRequest, toHex, Transaction } from 'viem'
+import { call, getTransaction, prepareTransactionRequest } from 'viem/actions'
 import { useConnections } from 'wagmi'
 
 import { SupportedChain } from '@app/constants/chains'
@@ -33,17 +24,9 @@ import {
   CreateQueryKey,
 } from '@app/types'
 import { getReadableError } from '@app/utils/errors'
+import { getAccessList } from '@app/utils/query/getAccessList'
 import { wagmiConfig } from '@app/utils/query/wagmi'
-import { CheckIsSafeAppReturnType } from '@app/utils/safe'
 import { hasParaConnection } from '@app/utils/utils'
-
-type AccessListResponse = {
-  accessList: {
-    address: Address
-    storageKeys: Hex[]
-  }[]
-  gasUsed: Hex
-}
 
 export const getUniqueTransaction = ({
   txKey,
@@ -120,47 +103,24 @@ export const registrationGasFeeModifier = (gasLimit: bigint, transactionName: Tr
 export const calculateGasLimit = async ({
   client,
   connectorClient,
-  isSafeApp,
   txWithZeroGas,
   transactionName,
 }: {
   client: ClientWithEns
   connectorClient: ConnectorClientWithEns
-  isSafeApp: boolean
   txWithZeroGas: BasicTransactionRequest
   transactionName: TransactionName
 }) => {
-  if (isSafeApp) {
-    const accessListResponse = await client.request<{
-      Method: 'eth_createAccessList'
-      Parameters: [tx: TransactionRequest<Hex>, blockTag: BlockTag]
-      ReturnType: AccessListResponse
-    }>({
-      method: 'eth_createAccessList',
-      params: [
-        {
-          to: txWithZeroGas.to,
-          data: txWithZeroGas.data,
-          from: connectorClient.account!.address,
-          value: toHex(txWithZeroGas.value ? txWithZeroGas.value + 1000000n : 0n),
-        },
-        'latest',
-      ],
-    })
-
-    return {
-      gasLimit: registrationGasFeeModifier(BigInt(accessListResponse.gasUsed), transactionName),
-      accessList: accessListResponse.accessList,
-    }
-  }
-
-  const gasEstimate = await estimateGas(client, {
-    ...txWithZeroGas,
-    account: connectorClient.account!,
+  const accessListResponse = await getAccessList(client, {
+    to: txWithZeroGas.to,
+    data: txWithZeroGas.data,
+    from: connectorClient.account!.address,
+    value: toHex(txWithZeroGas.value ? txWithZeroGas.value + 1000000n : 0n),
   })
+
   return {
-    gasLimit: registrationGasFeeModifier(gasEstimate, transactionName),
-    accessList: undefined,
+    gasLimit: registrationGasFeeModifier(BigInt(accessListResponse.gasUsed), transactionName),
+    accessList: accessListResponse.accessList,
   }
 }
 
@@ -195,7 +155,6 @@ export type CreateTransactionRequestQueryKey = CreateQueryKey<
 type CreateTransactionRequestUnsafeParameters = {
   client: ClientWithEns
   connectorClient: ConnectorClientWithEns
-  isSafeApp: CheckIsSafeAppReturnType | undefined
   params: UniqueTransaction
   chainId: SupportedChain['id']
   connections: any
@@ -204,7 +163,6 @@ type CreateTransactionRequestUnsafeParameters = {
 export const createTransactionRequestUnsafe = async ({
   client,
   connectorClient,
-  isSafeApp,
   params,
   chainId,
   connections,
@@ -225,7 +183,6 @@ export const createTransactionRequestUnsafe = async ({
   const { gasLimit, accessList } = await calculateGasLimit({
     client,
     connectorClient,
-    isSafeApp: !!isSafeApp,
     txWithZeroGas,
     transactionName: params.name,
   })
@@ -261,11 +218,9 @@ export const createTransactionRequestQueryFn =
   (config: ConfigWithEns) =>
   ({
     connectorClient,
-    isSafeApp,
     connections,
   }: {
     connectorClient: ConnectorClientWithEns | undefined
-    isSafeApp: CheckIsSafeAppReturnType | undefined
     connections: ReturnType<typeof useConnections>
   }) =>
   async ({
@@ -282,7 +237,6 @@ export const createTransactionRequestQueryFn =
         data: await createTransactionRequestUnsafe({
           client,
           connectorClient,
-          isSafeApp,
           params,
           chainId,
           connections,
