+++ src/hooks/chain/useEstimateGasWithStateOverride.ts
@@ -14,7 +14,6 @@ import {
   parseEther,
   RpcTransactionRequest,
   toHex,
-  TransactionRequest,
 } from 'viem'
 import { useConnectorClient } from 'wagmi'
 
@@ -31,8 +30,10 @@ import {
   Prettify,
   QueryConfig,
 } from '@app/types'
+import { emptyAddress } from '@app/utils/constants'
 import { getIsCachedData } from '@app/utils/getIsCachedData'
 import { prepareQueryOptions } from '@app/utils/prepareQueryOptions'
+import { getAccessList } from '@app/utils/query/getAccessList'
 import { useQuery } from '@app/utils/query/useQuery'
 
 import { useGasPrice } from './useGasPrice'
@@ -164,10 +165,28 @@ const estimateIndividualGas = async <TName extends TransactionName>({
     name,
   })
 
+  // SCAs use >21k gas to execute a transfer (most of the time, but depends on implementation)
+  // For any SCA transaction involving a transfer, the transaction will probably fail by default due to the extra gas.
+  // To fix this, we can use an access list of what storage slots are accessed in a transfer to pre-pay the cold gas,
+  // then when the transfer is done, only warm gas is used and the transfer is under the limit.
+  // Normally, you can just pull the estimated gas via `eth_getAccessList`, since it returns the access list + gas used,
+  // but since we're using a state override and that method doesn't support it, we have to do it manually.
+  // (Technically geth now has an implementation, but it's very new and only in geth, see https://github.com/ethereum/go-ethereum/issues/27630)
+  //
+  // To get the access list, we're executing the bytecode of this Yul code: https://gist.github.com/TateB/777287c9a63d5f02fcd905232ce5748a
+  // (note: 0xed3869F3020315C839b2f4E9a73bEbE9a9670534 is replaced with `connectorClient.account.address`)
+  // It does a simple transfer, and accesses any storage slot that would be accessed by any other transfer.
+  const accessList = await getAccessList(client, {
+    from: emptyAddress,
+    data: concatHex(['0x5f808080600173', connectorClient.account.address, '0x5af100']),
+    value: '0x1',
+  })
+
   const formattedRequest = formatTransactionRequest({
     ...generatedRequest,
     from: connectorClient.account.address,
-  } as TransactionRequest)
+    accessList: accessList.accessList,
+  })
 
   const stateOverrideWithBalance = stateOverride?.find(
     (s) => s.address === connectorClient.account.address,
